requirements.txt
# Kafka Dependencies
confluent-kafka==2.3.0
kafka-python==2.0.2

# Data Processing
pydantic==2.5.0
pyarrow==14.0.1
pandas==2.1.4

# Schema Validation
jsonschema==4.20.0
pyyaml==6.0.1
user-agents==2.2.0

# Data Generation
faker==20.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
python-dotenv==1.0.0
structlog==23.2.0

# Development
black==23.11.0
flake8==6.1.0 